+++ lib/ui/settings/pow/work_source_details_sheet.dart
@@ -145,7 +145,7 @@ class WorkSourceDetailsSheetState extends State<WorkSourceDetailsSheet> {
                         autocorrect: false,
                         keyboardType: TextInputType.text,
                         inputFormatters: [
-                          LengthLimitingTextInputFormatter(100),
+                          LengthLimitingTextInputFormatter(200),
                         ],
                         style: TextStyle(
                           fontWeight: FontWeight.w600,
@@ -163,7 +163,7 @@ class WorkSourceDetailsSheetState extends State<WorkSourceDetailsSheet> {
                         autocorrect: false,
                         keyboardType: TextInputType.text,
                         inputFormatters: [
-                          LengthLimitingTextInputFormatter(100),
+                          LengthLimitingTextInputFormatter(200),
                         ],
                         style: TextStyle(
                           fontWeight: FontWeight.w600,