+++ assets/CHANGELOG.md
@@ -1,9 +1,10 @@
-## v0.9.6 - [2024-03-01]
+## v0.9.6 - [2024-03-15]
 #### Changed
 - nanoauth / nanopay updates
 - Remove captcha option on gift cards
 - Updated a lot of dependencies
 - Tracking warning goes away permanently when dismissed now
+- Update URL input length limit 100 -> 200
 #### Fixed
 - Fix intro / welcome screen
 - Remove advanced options on the send confirm screen when they shouldn't be there